python/requirements.txt
pyspark>=3.3.0 
PyPDF2>=3.0.0
Pillow>=9.0.0
pytesseract>=0.3.0
pdf2image>=1.16.0